Move routes into their own folder
