+++ src/pages/index.ts
@@ -1,1 +1,5 @@
-export * from './AppRoutes';
+export * from './Forecast';
+export * from './Homepage';
+export * from './Login';
+export * from './Notes';
+export * from './Setup';
