--- src/pages/routes/index.ts
@@ -1,3 +1,0 @@
-export * from './ProtectedRoutes';
-export * from './PublicRoutes';
-export * from './AccountRoutes';
