+++ src/types/index.ts
@@ -2,6 +2,5 @@ export * from './OneOffPayment';
 export * from './account';
 export * from './actions';
 export * from './bill';
-export * from './modal';
 export * from './note';
 export * from './user';
