--- src/types/modal.ts
@@ -1,4 +1,0 @@
-export interface Modal {
-  PAYMENT_DUE: boolean;
-  BILL: boolean;
-}
